Add reprompt check when deleting ciphers in browser extension

+++ apps/browser/src/vault/popup/components/vault-v2/item-more-options/item-more-options.component.ts
@@ -267,6 +267,10 @@ export class ItemMoreOptionsComponent {
   }
 
   protected async delete() {
+    if (this.cipher.reprompt && !(await this.passwordRepromptService.showPasswordPrompt())) {
+      return;
+    }
+
     const confirmed = await this.dialogService.openSimpleDialog({
       title: { key: "deleteItem" },
       content: { key: "deleteItemConfirmation" },
